+++ lib/for_each.dart
@@ -4,7 +4,7 @@ import 'dart:convert';
 import 'package:cli_util/cli_logging.dart';
 import 'package:args/args.dart';
 // import 'package:args/src/utils.dart' show wrapText;
-import 'package:path/path.dart' as p;
+// import 'package:path/path.dart' as p;
 import 'package:glob/glob.dart';
 import 'package:glob/list_local_fs.dart';
 import 'package:for_each/file_info.dart';
@@ -12,19 +12,20 @@ import 'package:for_each/string_ext.dart';
 
 class ForEachApp {
   final String name = 'For-Each';
-  final String version = '1.2';
+  final String version = '1.3';
   final String copyright = 'Copyright (C) 2022 Andy Chentsov <chentsov@gmail.com>';
 
   String get intro => '$name $version, $copyright';
 
   static const int usageIndent = 2;
-  final String _usage = '''Usage: for-each [<options>] <mask> [in <path>] do <command...>
+  final String _usage = '''Usage: for-each [<options>] <mask> [in <path>] do [finally] <command...>
 
 You can use bash-style substitutions in the <mask>, for example: "assets/**/*.scss". 
 
 Substitutions can be used in <command>:
  - {path} for the full path to the found file;
  - {dir} for the full path to the found file without the file name itself;
+ - {parent-dir} for the path to the parent directory of the found file;
  - {file.ext} for the path to the found file relative to <path>;
  - {file} for the path (without extension) to the found file relative to <path>;
  - {name.ext} for the filename with extension;
@@ -32,6 +33,20 @@ Substitutions can be used in <command>:
  - {ext} for file extension.
 
 If the <path> is not specified, the current path is used.
+
+If the FINALLY keyword is specified before the command,
+then the command will be executed once for all found files.
+
+In the finally command, you can use variable substitution in
+the format "[text pattern]", inside "text pattern" you can use
+the substitutions described above for a regular command.
+Such patterns will be replicated for each found file.
+
+Example:
+> for-each *.* in example-project-folder do finally echo [Found: {name.ext};]
+
+If the files "file1.txt" and "file2.dat" are found, the following command will be executed:
+> echo Found: file1.txt; Found: file2.dat; 
 ''';
 
   String get usage => '''$_usage
@@ -46,6 +61,7 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
 
   int? failOn;
 
+  bool finallyRun = false;
   bool dryRun = false;
   bool silent = false;
   bool verbose = false;
@@ -80,6 +96,11 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
     verbosity = results['verbosity'] ?? verbosityNormal;
     logger = verbose ? Logger.verbose(logTime: false) : Logger.standard();
 
+    if (results['version']) {
+      displayVersion();
+      exitApp(98);
+    }
+
     if (rest.isEmpty) return;
 
     // mask
@@ -95,6 +116,12 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
 
     // command
     if (rest.first == 'do') {
+      rest.removeAt(0);
+    }
+
+    // finally option
+    if (rest.first == 'finally') {
+      finallyRun = true;
       rest.removeAt(0);
     }
 
@@ -118,10 +145,18 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
     exitApp(99);
   }
 
+  void displayVersion() {
+    logger.write('$version\n');
+  }
+
   void displaySummary() {
     logger.trace(' - Mask: $mask');
     logger.trace(' - Path: $path');
-    logger.trace(' - Command: $command ${commandArguments.join(' ')}');
+    if (finallyRun) {
+      logger.trace(' - Command (finally): $command ${commandArguments.join(' ')}');
+    } else {
+      logger.trace(' - Command: $command ${commandArguments.join(' ')}');
+    }
     logger.trace('');
   }
 
@@ -134,6 +169,7 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
       'dry-run',
       abbr: 'd',
       defaultsTo: false,
+      negatable: false,
       help: 'Dry run',
     );
 
@@ -141,6 +177,7 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
       'silent',
       abbr: 's',
       defaultsTo: false,
+      negatable: false,
       help: 'Silent output',
     );
 
@@ -148,6 +185,7 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
       'verbose',
       abbr: 'v',
       defaultsTo: false,
+      negatable: false,
       help: 'Verbose output',
     );
 
@@ -169,9 +207,17 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
     parser.addOption(
       'fail-on',
       abbr: 'f',
-      help: 'Exit with an <exit code> if the exit code of at least one command is greater than <exit code>',
+      help:
+          'Exit with an <exit code> if the exit code of at least one command is greater than <exit code>. Ignored for the finally command.',
       valueHelp: 'exit code',
       defaultsTo: null,
+    );
+
+    parser.addFlag(
+      'version',
+      defaultsTo: false,
+      negatable: false,
+      help: 'Prints the version of the $name',
     );
   }
 
@@ -197,18 +243,24 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
       logger.write('\n');
     }
 
+    var exitCode = 0;
+
     if (files.isEmpty) {
       if (!silent) {
         logger.write('No files found\n');
       }
     } else {
-      for (var file in files) {
-        final exitCode = await runBatch(file);
-        if (failOn != null && exitCode >= failOn!) {
-          if (!silent && progress != null) progress.finish(showTiming: true);
+      if (!finallyRun) {
+        for (var file in files) {
+          final _exitCode = await runBatch(file);
+          if (failOn != null && _exitCode >= failOn!) {
+            if (!silent && progress != null) progress.finish(showTiming: true);
 
-          exit(exitCode);
+            exit(_exitCode);
+          }
         }
+      } else {
+        exitCode = await runBatchFinally(files);
       }
     }
 
@@ -216,14 +268,16 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
 
     if (!silent) {
       logger.stdout('\nAll ${logger.ansi.emphasized('done')}.');
-      logger.flush();
     }
+
+    exit(exitCode);
   }
 
   String subst(String command, FileInfo file) {
     final result = command
         .replaceAll('{file}', file.basename)
         .replaceAll('{dir}', file.dir)
+        .replaceAll('{parent-dir}', file.parentPath)
         .replaceAll('{path}', file.path)
         .replaceAll('{file.ext}', file.name)
         .replaceAll('{ext}', file.ext)
@@ -265,6 +319,51 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
     }
   }
 
+  /// Substitute [pattern {var}]
+  ///
+  /// Example:
+  ///   sass [-I {dir}]
+  ///
+  String substList(String command, List<FileInfo> files) {
+    final result = command.replaceAllMapped(RegExp(r'\[(.*?)\]', caseSensitive: false), (match) {
+      return files.map((file) => subst(match.group(1)!, file)).join(' ');
+    });
+    return result;
+  }
+
+  Future<int> runBatchFinally(List<FileInfo> files) async {
+    final cmd = substList(command, files);
+    final args = commandArguments.map((entry) {
+      return substList(entry, files);
+    }).toList(growable: false);
+
+    if (dryRun) {
+      logger.write('RUN: $cmd ${args.join(' ')}\n');
+
+      return 0;
+    } else {
+      if (verbosity == verbosityDebug) {
+        logger.trace('RUN: $cmd ${args.join(' ')}\n');
+      }
+
+      final batchProcess = await Process.start(
+        cmd,
+        args,
+        runInShell: true,
+      );
+
+      unawaited(batchProcess.stdout.transform(utf8.decoder).forEach((s) {
+        logger.stdout(s.endsWith('\n') ? s.substring(0, s.length - 1) : s);
+      }));
+      unawaited(batchProcess.stderr.transform(utf8.decoder).forEach((s) {
+        logger.stderr(s.endsWith('\n') ? s.substring(0, s.length - 1) : s);
+      }));
+
+      final exitCode = await batchProcess.exitCode;
+      return exitCode;
+    }
+  }
+
   Future<List<FileInfo>> scan({
     required String mask,
     required String path,
@@ -280,16 +379,9 @@ ${_indent('${parser.usage}\n', indent: usageIndent)}
 
       // print(' * ${entry.path}');
 
-      final relativePath = p.relative(entry.path, from: path);
-
       return FileInfo(
-        path: entry.path,
-        dir: entry.dirname,
-        name: relativePath,
-        basename: p.withoutExtension(relativePath),
-        ext: p.extension(entry.basename).substring(1),
-        filename: p.withoutExtension(entry.basename),
-        filenameExt: entry.basename,
+        basePath: path,
+        entry: entry,
       );
     }).toList();