+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.3.0
+
+- Added the **finally** command run mode, which allows you to run the command once for all found files.
+- Added new substitution variable to the command: `{parent-dir}`.
+- Added `--version` option to print version number.
+
+## 1.2.0
+
+- Added new substitution variables to the command: `{name.ext}`, `{name}`.
+- Fixed substitution of variables in the command - now the variable can be used several times.
+
 ## 1.0.0
 
 - Initial version